WIP: change validation to an interface
emitCallsToValidators2 processes "Items" instead of Functions.  An item
can be a function or a group of functions or a comment, or...

This is just a POC to show the approach.

It's not clear what the API to validators should look like or whether we
can replace `Validations{}` with a single Validation or not.